pyproject.toml
[project]
name = "summary-paper-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "arxiv>=2.2.0",
    "vllm>=0.10.1.1",
    "slack-bolt>=1.18.0",
    "slack-sdk>=3.27.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "trafilatura>=2.0.0",
]
